Change the return type of Block's size-related methods to long
Currently the getSizeInBytes/getRegionSizeInBytes/getRetainedSizeInBytes
methods all return integers. To implement these methods Block implementations
usually sum up the size of their internal states and cast that sum to integer
using different cast functions, which are different in different Block implementations,
and occasionally incorrect (using Ints.saturatedCast()).

+++ presto-main/src/test/java/com/facebook/presto/block/AbstractTestBlock.java
@@ -187,17 +187,17 @@ public abstract class AbstractTestBlock
     {
         // Asserting on `block` is not very effective because most blocks passed to this method is compact.
         // Therefore, we split the `block` into two and assert again.
-        int expectedBlockSize = copyBlock(block).getSizeInBytes();
+        long expectedBlockSize = copyBlock(block).getSizeInBytes();
         assertEquals(block.getSizeInBytes(), expectedBlockSize);
         assertEquals(block.getRegionSizeInBytes(0, block.getPositionCount()), expectedBlockSize);
 
         List<Block> splitBlock = splitBlock(block, 2);
         Block firstHalf = splitBlock.get(0);
-        int expectedFirstHalfSize = copyBlock(firstHalf).getSizeInBytes();
+        long expectedFirstHalfSize = copyBlock(firstHalf).getSizeInBytes();
         assertEquals(firstHalf.getSizeInBytes(), expectedFirstHalfSize);
         assertEquals(block.getRegionSizeInBytes(0, firstHalf.getPositionCount()), expectedFirstHalfSize);
         Block secondHalf = splitBlock.get(1);
-        int expectedSecondHalfSize = copyBlock(secondHalf).getSizeInBytes();
+        long expectedSecondHalfSize = copyBlock(secondHalf).getSizeInBytes();
         assertEquals(secondHalf.getSizeInBytes(), expectedSecondHalfSize);
         assertEquals(block.getRegionSizeInBytes(firstHalf.getPositionCount(), secondHalf.getPositionCount()), expectedSecondHalfSize);
     }
