+++ presto-main/src/test/java/com/facebook/presto/block/TestVariableWidthBlock.java
@@ -95,11 +95,11 @@ public class TestVariableWidthBlock
         Block block = blockBuilder.build();
 
         List<Block> splitQuarter = splitBlock(block, 4);
-        int sizeInBytes = block.getSizeInBytes();
-        int quarter1size = splitQuarter.get(0).getSizeInBytes();
-        int quarter2size = splitQuarter.get(1).getSizeInBytes();
-        int quarter3size = splitQuarter.get(2).getSizeInBytes();
-        int quarter4size = splitQuarter.get(3).getSizeInBytes();
+        long sizeInBytes = block.getSizeInBytes();
+        long quarter1size = splitQuarter.get(0).getSizeInBytes();
+        long quarter2size = splitQuarter.get(1).getSizeInBytes();
+        long quarter3size = splitQuarter.get(2).getSizeInBytes();
+        long quarter4size = splitQuarter.get(3).getSizeInBytes();
         double expectedQuarterSizeMin = sizeInBytes * 0.2;
         double expectedQuarterSizeMax = sizeInBytes * 0.3;
         assertTrue(quarter1size > expectedQuarterSizeMin && quarter1size < expectedQuarterSizeMax, format("quarter1size is %s, should be between %s and %s", quarter1size, expectedQuarterSizeMin, expectedQuarterSizeMax));
