+++ presto-spi/src/main/java/com/facebook/presto/spi/block/AbstractInterleavedBlock.java
@@ -253,7 +253,7 @@ public abstract class AbstractInterleavedBlock
     }
 
     @Override
-    public int getRegionSizeInBytes(int position, int length)
+    public long getRegionSizeInBytes(int position, int length)
     {
         if (position == 0 && length == getPositionCount()) {
             // Calculation of getRegionSizeInBytes is expensive in this class.
